Moved utilities file to root folder
